Added spinner, styles were made in a separate file

+++ EngManVue/src/components/Messages.vue
@@ -1,10 +1,11 @@
 <template>
   <div class = 'messages-view'>
-      <div class = "label-form">
+      <div class="loading" v-if = "inProgress">Loading&#8230;</div>
+      <div class = "label-form-mes">
             <span>Chat</span>
             <span style = "margin-left: 370px; font-size:10px; cursor: pointer" v-on:click = "closeform()">close</span>
       </div>
-      <select class = "select-form" v-model = "beneficiaryEmail" v-on:change = "changeBeneficiary(beneficiaryEmail)">
+      <select class = "select-form-mes" v-model = "beneficiaryEmail" v-on:change = "changeBeneficiary(beneficiaryEmail)">
             <option v-for = "user in users" v-if = "user.Id != $store.state.user.Id" :key = "user.Id">
                 {{user.Email}}
             </option>
@@ -16,7 +17,7 @@
           <div v-for = "el in sortedmessages" :key = "el.MessageId">
               <div v-if = "el.Sender.Id == sender.Id" class = "label-sender"><span style = "font-size: 10px; cursor: pointer" v-on:click = "deleteMessage(el.MessageId)">delete</span> You</div>
               <div v-if = "el.Sender.Id == sender.Id" class = "message-sender">
-                  {{el.Text}}
+                  {{el.Text}} - {{dateTime(el.Time)}}
               </div>
               <div v-if = "el.Sender.Id != sender.Id" class = "label-beneficiary">{{el.Sender.FirstName}} <span style = "font-size: 10px; cursor: pointer" v-on:click = "deleteMessage(el.MessageId)">delete</span></div>
               <div v-if = "el.Sender.Id != sender.Id" class = "message-beneficiary">
@@ -24,8 +25,8 @@
               </div>
           </div>
       </div>
-      <div class = "input-form">
-            <textarea type = "text" v-model = "message" :disabled = "beneficiaryEmail == 'none'"/>
+      <div class = "input-form-mes">
+            <textarea class = "textarea-mes" type = "text" v-model = "message" :disabled = "beneficiaryEmail == 'none'"/>
       </div>
       <button v-on:click = "sendMessage">Отправить</button>
   </div>
@@ -38,6 +39,7 @@ export default {
   name: 'messages-view',
   data: () => {
       return {
+          inProgress: false,
           message: '',
           beneficiary: undefined,
           beneficiaryEmail: 'none',
@@ -46,6 +48,7 @@ export default {
   },
   methods: {
       closeform(){
+          this.inProgress = false;
           this.message = '';
           this.beneficiary = undefined;
           this.beneficiaryEmail = 'none';
@@ -53,35 +56,52 @@ export default {
           this.$emit('closeMessages');
       },
       deleteMessage(id){
+          if(this.inProgress) return;
+          this.inProgress = true;
           if(confirm("Вы уверены, что хотите удалить сообщение?") == true){
             api.deleteMessage(id)
             .then(result =>{
                 this.$store.dispatch('getMessages');
+                this.inProgress = false;
             })
           } else{
               alert("Вы отменили удаление!");
+              this.inProgress = false;
           }
       },
       changeBeneficiary(email){
+          if(this.inProgress) return;
+          this.inProgress = true;
           this.message = '';
           if(this.beneficiaryEmail == 'none'){
                this.beneficiary = undefined;
+               this.inProgress = false;
                return;
           }
           this.beneficiary = this.$store.getters.users.filter(function(user){
               return user.Email == email;
           })[0];
+          this.inProgress = false;
       },
       sendMessage(){
+          if(this.inProgress) return;
+          this.inProgress = true;
           if(this.beneficiary != undefined)
           {
             if(this.message.length > 0)
             {
+                var date = new Date(Date.now());
+                var hours = date.getHours() < 10 ? "0"+date.getHours() : date.getHours();
+                var minutes = date.getMinutes() < 10 ? "0"+date.getMinutes() : date.getMinutes();
+                var seconds = date.getSeconds() < 10 ? "0"+date.getSeconds() : date.getSeconds();
+                var month = (date.getMonth()+1) < 10 ? "0"+(date.getMonth()+1) : (date.getMonth()+1);
+                var day = date.getDate < 10 ? "0"+date.getDate() : date.getDate();
                 api.sendMessage({
                     MessageId: 0,
                     SenderId: this.sender.Id,
                     BeneficiaryId: this.beneficiary.Id,
-                    Text: this.message
+                    Text: this.message,
+                    Time: date.getFullYear() + "-" + month + "-" + day + "T" + hours + ":" + minutes + ":" + seconds + "+03:00"
                 })
                 .then(res => 
                 {
@@ -93,6 +113,13 @@ export default {
             }   
           }
           this.message = '';
+          this.inProgress = false;
+      },
+      dateTime(_date){
+                var date = new Date(_date);
+                var hours = date.getHours() < 10 ? "0"+date.getHours() : date.getHours();
+                var minutes = date.getMinutes() < 10 ? "0"+date.getMinutes() : date.getMinutes();
+                return hours + ":" + minutes;
       }
   },
   computed: {
@@ -114,92 +141,3 @@ export default {
   }
 }
 </script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-.messages-view{
-    position: fixed;
-    right: 1%;
-    bottom: 2.5%;
-    width: 450px;
-    height: 500px;
-    text-align: center;
-    display: table;
-    background: rgb(248, 248, 248);
-    border: solid 1px;
-    outline:none;
-    border-radius: 10px;
-}
-.messages{
-    margin: 0.5%;
-    margin-left: 10px;
-    height: 345px;
-    overflow-y: scroll;
-}
-.label-form{
-    background: rgba(0,0,0,.15);
-    border-radius: 10px 10px 0 0;
-    text-align: left;
-    padding: 0.5%;
-    height: 25px;
-    font-weight: bold;
-}
-.message-beneficiary{
-    margin: 0.5%;
-    text-align: left;
-    padding: 0.5%;
-    border: none;
-    border-bottom-style: dashed;
-    border-width: 0.7px;
-    outline:none;
-    font-size: 16px;
-}
-.message-sender{
-    margin: 0.5%;
-    text-align: right;
-    padding: 0.5%;
-    border: none;
-    border-bottom-style: dashed;
-    border-width: 0.7px;
-    outline:none;
-    font-size: 16px;
-}
-.label-sender{
-    margin: 0.5%;
-    text-align: right;
-    font-size: 16px;
-    font-weight: bold;
-}
-.label-beneficiary{
-    margin: 0.5%;
-    text-align: left;
-    font-size: 16px;
-    font-weight: bold;
-}
-.input-form{
-    height: 50px;
-}
-textarea{
-    width: 400px;
-    height: 45px;
-    border-radius: 10px;
-    border-style: solid;
-    border-width: 1px;
-    border-color:rgb(75, 75, 75);
-    outline: none;
-    resize: none;
-    padding: 0.5%;
-}
-.select-form{
-    outline:none;
-    background: rgba(0,0,0,.1);
-    border: none;
-    border-radius: 10px;
-    text-align: center;
-    font-size: 16px;
-    margin: 0.5%;
-    resize: none;
-    height: 25px;
-    width: 350px;
-}
-</style>
